refactor(neural-link): migrate services to flat SDK boundary (#10994)
Migrates 9 entities from ai/mcp/server/neural-link/services/ to the
canonical ai/services/neural-link/ flat SDK boundary, mirroring the
KB (#10995) and GH-WF (#10997) M6 patterns. Updates ai/services.mjs
alias imports, NL Server.mjs custom boot() override service-imports,
and 1 test spec relocated to canonical SDK location.

Scope:
- 9 services: ComponentService, ConnectionService, DataService,
  HealthService, InstanceService, InteractionService, RecorderService,
  RuntimeService, toolService
- All 7 currently-aliased NeuralLink_* services preserved at
  ai/services.mjs (consumer-side contract unchanged); RecorderService
  + toolService remain internal
- Internal cross-references updated for new relative paths:
  - ../config.mjs → ../../mcp/server/neural-link/config.mjs
  - ../logger.mjs → ../../mcp/server/neural-link/logger.mjs
  - ../../../../../src/core/Base.mjs → ../../../src/core/Base.mjs
- toolService.mjs special-case import path fixes:
  - '../../../ToolService.mjs' → '../../mcp/ToolService.mjs'
    (resolves to ai/mcp/ToolService.mjs)
  - '../Server.mjs' (NL Server.mjs's getCurrentTurnId export per
    #10975) → '../../mcp/server/neural-link/Server.mjs'
  - path.join(__dirname, '../openapi.yaml') →
    '../../mcp/server/neural-link/openapi.yaml'
- NL Server.mjs custom boot() override imports updated for new SDK
  paths (transport-before-services semantic preserved per AC11)
- 1 unit spec (RecorderService.spec.mjs) moved from
  test/playwright/unit/ai/mcp/server/neural-link/ to
  test/playwright/unit/ai/services/neural-link/ matching new SDK shape
- Test path math: 7→6 ../ for src+ai imports + 5→4 ../ for setup.mjs
- logger.spec.mjs retained at original location (server-level, not
  service)

Custom-boot-order preservation per #10975 + #10994 AC11: NL Server.mjs
boot() override transport-before-services semantic + bridgeCwd member +
_turnId counter + getCurrentTurnId() export all preserved post-migration.
