+++ ai/services/neural-link/ComponentService.mjs
@@ -1,4 +1,4 @@
-import Base              from '../../../../../src/core/Base.mjs';
+import Base              from '../../../src/core/Base.mjs';
 import ConnectionService from './ConnectionService.mjs';
 
 /**
