+++ ai/services/neural-link/InteractionService.mjs
@@ -1,4 +1,4 @@
-import Base              from '../../../../../src/core/Base.mjs';
+import Base              from '../../../src/core/Base.mjs';
 import ConnectionService from './ConnectionService.mjs';
 
 /**
